Add files via upload
Added options -L (with layout; default) and -l (without layout). When calling lily2data with -l, score is not printed and only the midi file is created.
Now permusic.sh stops when one of the production stages fail.
New script permusic_nopdf.sh is the same as permusic.h but option -l is issued for lily2data (see above).